refactor(server): isolate authority conflict resolution
